Split Faraday middleware
Split into two middlewares, so separate Acme features that are specific only to the Acme connections.
