+++ spec/client_http_client_spec.rb
@@ -1,0 +1,4 @@
+require 'spec_helper'
+
+describe Acme::Client::HTTPClient do
+end
